Extract exercise list into component for reuse
